chore: reorganize components/dashboard folder
